requirements.txt
python-telegram-bot==20.3
openai
gspread
oauth2client